+++ src/Components/Svg.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace BladeUI\Icons\Components;
 
+use Closure;
 use Illuminate\View\Component;
 
 final class Svg extends Component
 {
-    public function render()
+    public function render(): Closure
     {
         return function (array $data) {
             $attributes = $data['attributes']->getIterator()->getArrayCopy();